Moved blog methods into the new AdminController
