+++ src/Controller/FrontController.php
@@ -28,10 +28,6 @@ class FrontController extends Controller
      */
     protected $twig = null;
     /**
-     * @var null
-     */
-    protected $page = null;
-    /**
      * @var string
      */
     protected $controller = self::DEFAULT_CONTROLLER;
