+++ src/View/admin.twig
@@ -51,10 +51,11 @@
                                         : </b><em>{{ post.add_date | date("d/m/Y") }}</em><br>
                                     <div class="mt-1">
                                         <a class="btn btn-sm btn-outline-dark"
-                                           href="../public/?id={{ post.id }}&access=blog!edit">Editer <i
+                                           href="../public/?id={{ post.id }}&access=admin!edit">Editer <i
                                                     class="far fa-edit"></i></a>
                                         <a class="btn btn-sm btn-outline-danger"
-                                           href="../public/?id={{ post.id }}&table=post&access=blog!confirm">Supprimer <i
+                                           href="../public/?id={{ post.id }}&table=post&access=admin!confirm">Supprimer
+                                            <i
                                                     class="fas fa-trash"></i></a></div>
                                 </td>
                             </tr>
@@ -68,10 +69,11 @@
                             <p class="h5 font-weight-bold mb-3">Vous êtes sur le point de supprimer cet article ainsi
                                 que
                                 ses commentaires !<br>Souhaitez-vous continuer ?</p>
-                            <a class="btn btn-outline-danger" href="../public/?id={{ confirm.id }}&table=post&access=blog!delete">Confirmer <i
+                            <a class="btn btn-outline-danger"
+                               href="../public/?id={{ confirm.id }}&table=post&access=admin!delete">Confirmer <i
                                         class="fas fa-trash"></i></a>
                             <a class="btn btn-outline-success"
-                               href="../public/?table=cancel&show=0&access=blog!confirm">Annuler</a>
+                               href="../public/?table=cancel&show=0&access=admin!confirm">Annuler</a>
                         </div>
                     {% endif %}
                 </div>
@@ -98,12 +100,13 @@
                                     <div class="mt-1">
                                         {% if comment.validation == 0 %}
                                             <a class="btn btn-sm btn-outline-dark"
-                                               href="../public/?id={{ comment.id }}&access=blog!validate">
+                                               href="../public/?id={{ comment.id }}&access=admin!validate">
                                                 Valider <i
                                                         class="fas fa-check"></i></a>
                                         {% endif %}
                                         <a class="btn btn-sm btn-outline-danger"
-                                           href="../public/?id={{ comment.id }}&table=comment&access=blog!confirm">Supprimer <i
+                                           href="../public/?id={{ comment.id }}&table=comment&access=admin!confirm">Supprimer
+                                            <i
                                                     class="fas fa-trash"></i></a>
                                     </div>
                                 </td>
@@ -116,10 +119,11 @@
                         <div class="text-center mt-5">
                             <p class="h4 font-weight-bolder">Confirmation</p><br>
                             <p class="h5 font-weight-bold mb-3">Souhaitez-vous supprimer ce commentaire ?</p>
-                            <a class="btn btn-outline-danger" href="../public/?id={{ confirm.id }}&table=comment&access=blog!delete">Confirmer <i
+                            <a class="btn btn-outline-danger"
+                               href="../public/?id={{ confirm.id }}&table=comment&access=admin!delete">Confirmer <i
                                         class="fas fa-trash"></i></a>
                             <a class="btn btn-outline-success"
-                               href="../public/?table=cancel&show=1&access=blog!confirm">Annuler</a>
+                               href="../public/?table=cancel&show=1&access=admin!confirm">Annuler</a>
                         </div>
                     {% endif %}
                 </div>
@@ -143,7 +147,8 @@
                                     <b>Email : </b>:{{ user.email }}
                                     <div class="mt-1">
                                         <a class="btn btn-sm btn-outline-danger"
-                                           href="../public/?id={{ user.id }}&table=user&access=blog!confirm">Supprimer <i
+                                           href="../public/?id={{ user.id }}&table=user&access=admin!confirm">Supprimer
+                                            <i
                                                     class="fas fa-trash"></i></a>
                                     </div>
                                 </td>
@@ -158,10 +163,10 @@
                             <p class="h5 font-weight-bold mb-3">Vous êtes sur le point de supprimer cet utilisateur.<br>Souhaitez-vous
                                 continuer ?</p>
                             <a class="btn btn-outline-danger"
-                               href="../public/?id={{ confirm.id }}&table=user&access=blog!delete">Confirmer <i
+                               href="../public/?id={{ confirm.id }}&table=user&access=admin!delete">Confirmer <i
                                         class="fas fa-trash"></i></a>
                             <a class="btn btn-outline-success"
-                               href="../public/?table=cancel&show=2&access=blog!confirm">Annuler</a>
+                               href="../public/?table=cancel&show=2&access=admin!confirm">Annuler</a>
                         </div>
                     {% endif %}
                 </div>
@@ -170,7 +175,7 @@
                  aria-labelledby="pills-newpost-tab">
 
                 <form name="sentMessage" id="editForm" enctype="multipart/form-data" method="post"
-                      action="../public/?access=blog!add">
+                      action="../public/?access=admin!add">
                     <div class="has-success">
                         <div class="col-lg-8 col-md-10 mx-auto mt-2">
                             <h3 class="mb-3 border-bottom text-center col-12">Edition</h3>
