+++ src/View/edit.twig
@@ -6,7 +6,7 @@
             <div class="col-lg-8 col-md-10 mx-auto mt-5">
                 <h3 class="mb-3 border-bottom text-center col-12">Edition</h3>
                 <form name="sentMessage" id="editForm" method="post" enctype="multipart/form-data"
-                      action="../public/?id={{ post.id }}&access=blog!update">
+                      action="../public/?id={{ post.id }}&access=admin!update">
                     <div class="control-group">
                         <div class="form-group floating-label-form-group controls">
                             <label class="font-weight-bold" for="author">Auteur</label>
@@ -46,7 +46,7 @@
                 </form>
                 <!-- Pager -->
                 <div class="clearfix mb-5">
-                    <a class="btn btn-outline-dark float-right" href="../public/?p=404&access=blog">Retour &rarr;</a>
+                    <a class="btn btn-outline-dark float-right" href="../public/?access=admin">Retour &rarr;</a>
                 </div>
             </div>
         </div>
