+++ src/View/layout/header.twig
@@ -24,7 +24,7 @@
             {% endif %}
             {% if session.user.status == true %}
                 <li class="nav-item">
-                    <a class="nav-link" href="../public/?p=404&access=blog">Administration</a>
+                    <a class="nav-link" href="../public/?access=admin">Administration</a>
                 </li>
             {% endif %}
         </ul>
